refactor: tidy up a few things
