+++ .github/workflows/release.yml
@@ -1,120 +1,34 @@
 name: Release
 
+# This GitHub action creates a release when a tag that matches the pattern
+# "v*" (e.g. v0.1.0) is created.
 on:
   push:
     tags:
-      - "v*"
-  workflow_dispatch:
-    inputs:
-      tag:
-        description: "Tag to release (e.g. v1.2.3). Required when manually running."
-        required: true
-        type: string
+      - 'v*'
 
+# Releases need permissions to read and write the repository contents.
+# GitHub considers creating releases and uploading assets as writing contents.
 permissions:
   contents: write
 
-concurrency:
-  group: release-${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: false
-
 jobs:
-  build:
-    name: Build (${{ matrix.goos }}/${{ matrix.goarch }})
+  goreleaser:
     runs-on: ubuntu-latest
-
-    strategy:
-      fail-fast: false
-      matrix:
-        goos: [linux, darwin, windows]
-        goarch: [amd64, arm64]
-
     steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Set up Go
-        uses: actions/setup-go@v5
+      - uses: actions/checkout@692973e3d937129bcbf40652eb9f2f61becf3332 # v4.1.7
         with:
-          go-version-file: go.mod
+          # Allow goreleaser to access older tag information.
+          fetch-depth: 0
+      - uses: actions/setup-go@0a12ed9d6a96ab950c8f026ed9f722fe0da7ef32 # v5.0.2
+        with:
+          go-version-file: 'go.mod'
           cache: true
 
-      - name: Unit tests
-        run: go test ./...
-
-      - name: Build
-        shell: bash
-        run: |
-          set -euo pipefail
-
-          version="${GITHUB_REF_NAME:-}"
-          if [[ -z "$version" || "$version" == "refs/tags/"* ]]; then
-            version="${{ inputs.tag || '' }}"
-          fi
-          if [[ -z "$version" ]]; then
-            echo "Unable to determine version/tag" >&2
-            exit 1
-          fi
-
-          ext=""
-          if [[ "${{ matrix.goos }}" == "windows" ]]; then
-            ext=".exe"
-          fi
-
-          out_dir="dist"
-          mkdir -p "$out_dir"
-
-          bin_name="tfmodmake_${version}_${{ matrix.goos }}_${{ matrix.goarch }}${ext}"
-          bin_path="$out_dir/$bin_name"
-
-          CGO_ENABLED=0 GOOS="${{ matrix.goos }}" GOARCH="${{ matrix.goarch }}" \
-            go build -trimpath -ldflags "-s -w" -o "$bin_path" ./cmd/tfmodmake
-
-          if [[ "${{ matrix.goos }}" == "windows" ]]; then
-            pkg_name="tfmodmake_${version}_${{ matrix.goos }}_${{ matrix.goarch }}.zip"
-            (cd "$out_dir" && zip -9 "$pkg_name" "$bin_name" >/dev/null)
-            rm -f "$bin_path"
-          else
-            pkg_name="tfmodmake_${version}_${{ matrix.goos }}_${{ matrix.goarch }}.tar.gz"
-            (cd "$out_dir" && tar -czf "$pkg_name" "$bin_name")
-            rm -f "$bin_path"
-          fi
-
-          echo "Built $out_dir/$pkg_name"
-
-      - name: Upload artifact
-        uses: actions/upload-artifact@v4
+      - name: Run GoReleaser
+        uses: goreleaser/goreleaser-action@286f3b13b1b49da4ac219696163fb8c1c93e1200 # v6.0.0
         with:
-          name: dist-${{ matrix.goos }}-${{ matrix.goarch }}
-          path: dist/*
-          if-no-files-found: error
-
-  release:
-    name: Create GitHub Release
-    runs-on: ubuntu-latest
-    needs: build
-
-    steps:
-      - name: Download artifacts
-        uses: actions/download-artifact@v4
-        with:
-          pattern: dist-*
-          merge-multiple: true
-          path: dist
-
-      - name: Generate checksums
-        shell: bash
-        run: |
-          set -euo pipefail
-          cd dist
-          sha256sum * > checksums.sha256
-          echo "Generated dist/checksums.sha256"
-
-      - name: Create release
-        uses: softprops/action-gh-release@v2
-        with:
-          tag_name: ${{ github.ref_name || inputs.tag }}
-          name: ${{ github.ref_name || inputs.tag }}
-          files: |
-            dist/*
-          generate_release_notes: true
+          args: release --clean
+        env:
+          # GitHub sets the GITHUB_TOKEN secret automatically.
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
